requirements.txt
numpy
python-binance
pandas
requests
empyrical
statsmodels
plotly
matplotlib
seaborn
scipy
catboost
xgboost
shap